requirements.txt
absl-py @ file:///C:/b/abs_5babsu7y5x/croot/absl-py_1666362945682/work
aiohttp @ file:///C:/b/abs_c4zmy2l696/croot/aiohttp_1670009573673/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
anyio @ file:///C:/ci/anyio_1644481856696/work/dist
appdirs==1.4.4
argon2-cffi @ file:///C:/ci_310/argon2-cffi_1642107167485/work
asgiref==3.8.1
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
async-timeout @ file:///C:/b/abs_43ozhz2a8g/croots/recipe/async-timeout_1664876362767/work
attrs @ file:///C:/b/abs_09s3y775ra/croot/attrs_1668696195628/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///C:/b/abs_0agyz1wsr4/croot/beautifulsoup4-split_1681493048687/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
blinker==1.4
Bottleneck @ file:///C:/Windows/Temp/abs_3198ca53-903d-42fd-87b4-03e6d03a8381yfwsuve8/croots/recipe/bottleneck_1657175565403/work
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi @ file:///C:/b/abs_36eb5mzhph/croot/certifi_1690232276943/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///C:/ci/click_1646056762388/work
cmake==3.26.4
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
comm @ file:///C:/b/abs_1419earm7u/croot/comm_1671231131638/work
contourpy @ file:///C:/b/abs_d5rpy288vc/croots/recipe/contourpy_1663827418189/work
cryptography @ file:///C:/b/abs_b7d7drzbky/croot/cryptography_1673298763653/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
datetime-truncate==1.1.1
debugpy @ file:///C:/ci_310/debugpy_1642079916595/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
dj-database-url==2.1.0
Django==5.0.6
django-admin-charts==1.3.1
django-admin-datta==1.0.10
django-admin-soft-dashboard==1.0.12
django-bootstrap-modal-forms==2.2.0
django-bootstrap3==22.2
django-bower==5.2.0
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
django-light==0.1.0.post3
django-memoize==2.3.1
django-mssql-backend==2.8.1
django-multiselectfield==0.1.12
django-nvd3==0.10.1
django-widget-tweaks==1.4.12
face-recognition==1.3.0
face-recognition-models==0.3.0
fonttools==4.25.0
gunicorn==22.0.0
mkl-service==2.4.0
mssql-django==1.5
munkres==1.1.4
mysqlclient==2.2.4
Pillow==9.4.0
pyserial==3.5
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2022.7.1
pywin32==305.1
requests-oauthlib==1.3.0
scipy==1.10.0
sip==4.19.13
smbus2==0.4.3
sqlparse==0.4.3
wincertstore==0.2
